Advanced Relationships with Entity Framework Core

+++ Services/CharacterService/CharacterService.cs
@@ -1,11 +1,13 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using dotnet_rpg.Data;
 using dotnet_rpg.Dtos.Character;
 using dotnet_rpg.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 
 namespace dotnet_rpg.Services.CharacterService
@@ -14,29 +16,37 @@ namespace dotnet_rpg.Services.CharacterService
   {
     private readonly IMapper _mapper;
     private readonly DataContext _context;
-    
-    public CharacterService(IMapper mapper, DataContext context)
+    private readonly IHttpContextAccessor _httpContextAccessor;
+
+    public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
     {
         _mapper = mapper;
         _context = context;
+        _httpContextAccessor = httpContextAccessor;
     }
+
+    private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
 
     public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
     {
         var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
         var character = _mapper.Map<Character>(newCharacter);
+        character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
 
         await _context.Characters.AddAsync(character);
         await _context.SaveChangesAsync();
-        serviceResponse.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+        serviceResponse.Data = _context.Characters.Where(c => c.User.Id == GetUserId()).Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
 
         return serviceResponse;
     }
 
-    public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters(int userId)
+    public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
     {
         var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-        var dbCharacters = await _context.Characters.Where(c => c.User.Id == userId).ToListAsync();
+        var dbCharacters = await _context.Characters
+            .Include(c => c.Weapon)
+            .Include(c => c.CharacterSkills).ThenInclude(cs => cs.Skill)
+            .Where(c => c.User.Id == GetUserId()).ToListAsync();
         serviceResponse.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
 
         return serviceResponse;
@@ -45,7 +55,10 @@ namespace dotnet_rpg.Services.CharacterService
     public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
     {
         var serviceResponse = new ServiceResponse<GetCharacterDto>();
-        var dbCharacter = await _context.Characters.FirstOrDefaultAsync(x => x.Id == id);
+        var dbCharacter = await _context.Characters
+            .Include(c => c.Weapon)
+            .Include(c => c.CharacterSkills).ThenInclude(cs => cs.Skill)
+            .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
         serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
 
         return serviceResponse;
@@ -57,19 +70,27 @@ namespace dotnet_rpg.Services.CharacterService
 
         try
         {
-            var dbCharacter = await _context.Characters.FirstOrDefaultAsync(x => x.Id == updatedCharacter.Id);
-            var character = _mapper.Map<Character>(dbCharacter);
-            character.Class = updatedCharacter.Class;
-            character.Defense = updatedCharacter.Defense;
-            character.HitPoints = updatedCharacter.HitPoints;
-            character.Intelligence = updatedCharacter.Intelligence;
-            character.Name = updatedCharacter.Name;
-            character.Strength = updatedCharacter.Strength;
+            var character = await _context.Characters.Include(c => c.User).FirstOrDefaultAsync(x => x.Id == updatedCharacter.Id);
 
-            _context.Characters.Update(character);
-            await _context.SaveChangesAsync();
+            if(character.User.Id == GetUserId())
+            {
+                character.Class = updatedCharacter.Class;
+                character.Defense = updatedCharacter.Defense;
+                character.HitPoints = updatedCharacter.HitPoints;
+                character.Intelligence = updatedCharacter.Intelligence;
+                character.Name = updatedCharacter.Name;
+                character.Strength = updatedCharacter.Strength;
 
-            serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+                _context.Characters.Update(character);
+                await _context.SaveChangesAsync();
+
+                serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+            }
+            else
+            {
+                serviceResponse.Success = false;
+                serviceResponse.Message = "Character not found.";
+            }
         }
         catch(Exception ex)
         {
@@ -87,11 +108,18 @@ namespace dotnet_rpg.Services.CharacterService
 
         try
         {
-            var dbCharacter = await _context.Characters.FirstAsync(x => x.Id == id);
-            _context.Characters.Remove(dbCharacter);
-            await _context.SaveChangesAsync();
-
-            serviceResponse.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+            var dbCharacter = await _context.Characters.FirstAsync(c => c.Id == id && c.User.Id == GetUserId());
+            if(dbCharacter != null)
+            {
+                _context.Characters.Remove(dbCharacter);
+                await _context.SaveChangesAsync();
+                serviceResponse.Data = _context.Characters.Where(c => c.User.Id == GetUserId()).Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+            }
+            else
+            {
+                serviceResponse.Success = false;
+                serviceResponse.Message = "Character not found.";
+            }
         }
         catch (Exception ex)
         {
